refactor: reverted back to earlier sign in/sign up structure
I had completely missed that the sign in component was used in cart page, so I put the relevant form stuff into a component again
