+++ client/src/main.tsx
@@ -25,7 +25,7 @@ import { Faq } from './pages/Faq';
 import Home from './pages/Home';
 import ProductDetails from './pages/ProductDetails';
 import Shop from './pages/Shop';
-import { SignInForm } from './pages/SignIn';
+import { SignIn } from './pages/SignIn';
 import { default as SignUp } from './pages/SignUp';
 import UserOrders from './pages/UserOrders';
 import Admin from './pages/admin/Admin';
@@ -45,7 +45,7 @@ const router = createBrowserRouter(
         <Route path="product/:id" element={<ProductDetails />} />
         <Route path="checkout" element={<Cart />} />
         <Route path="confirmation" element={<Confirmation />} />
-        <Route path="signin" element={<SignInForm />} />
+        <Route path="signin" element={<SignIn />} />
         <Route path="signup" element={<SignUp />} />
         <Route path="orders" element={<UserOrders />} />
       </Route>
